Move auth in their own folder
