+++ src/auth/GenericTemplateAuthorization.ts
