+++ src/auth/QuestionaryAuthorization.ts
@@ -10,9 +10,9 @@ import { QuestionaryDataSource } from '../datasources/QuestionaryDataSource';
 import { SampleDataSource } from '../datasources/SampleDataSource';
 import { ShipmentDataSource } from '../datasources/ShipmentDataSource';
 import { TemplateDataSource } from '../datasources/TemplateDataSource';
+import { VisitDataSource } from '../datasources/VisitDataSource';
 import { TemplateGroupId } from '../models/Template';
 import { User, UserWithRole } from '../models/User';
-import { VisitDataSource } from './../datasources/VisitDataSource';
 import { EsiAuthorization } from './EsiAuthorization';
 import { UserAuthorization } from './UserAuthorization';
 import { VisitAuthorization } from './VisitAuthorization';
