+++ src/decorators/Authorized.ts
@@ -1,9 +1,9 @@
 import { container } from 'tsyringe';
 
+import { UserAuthorization } from '../auth/UserAuthorization';
 import { Rejection, rejection } from '../models/Rejection';
 import { Roles } from '../models/Role';
 import { UserWithRole } from '../models/User';
-import { UserAuthorization } from '../utils/UserAuthorization';
 
 const Authorized = (roles: Roles[] = []) => {
   return (
