+++ src/mutations/UserMutations.ts
@@ -14,6 +14,7 @@ import {
 import * as bcrypt from 'bcryptjs';
 import { container, inject, injectable } from 'tsyringe';
 
+import { UserAuthorization } from '../auth/UserAuthorization';
 import { Tokens } from '../config/Tokens';
 import UOWSSoapClient from '../datasources/stfc/UOWSSoapInterface';
 import { UserDataSource } from '../datasources/UserDataSource';
@@ -41,7 +42,6 @@ import {
   UpdateUserRolesArgs,
 } from '../resolvers/mutations/UpdateUserMutation';
 import { signToken, verifyToken } from '../utils/jwt';
-import { UserAuthorization } from '../utils/UserAuthorization';
 
 @injectable()
 export default class UserMutations {
